feat(result): add result page UI with breakdown and explanations
